server/requirements.txt
# ===================================
# MeetingApp Backend Dependencies
# ===================================

# Core Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Database
sqlalchemy==2.0.23
pyodbc==5.0.1

# Data Validation & Settings
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Authentication & Security
python-jose[cryptography]==3.3.0
bcrypt==4.0.1

# File Upload & Environment
python-multipart==0.0.6
python-dotenv==1.0.0

# OpenAI Integration
openai>=1.12.0
httpx>=0.24.0,<1.0.0

# Whisper Speech-to-Text
openai-whisper==20231117

# WebSocket Support
websockets==12.0

# Async File Operations
aiofiles==23.2.1
aiohttp>=3.9.0

# ===================================
# Audio Processing & Noise Reduction
# ===================================
librosa>=0.10.1
soundfile>=0.12.1
noisereduce>=3.0.0
numpy>=1.24.0,<2.0.0
scipy>=1.11.0

# ===================================
# Speaker Diarization (Pyannote)
# ===================================
pyannote.audio>=3.1.0

# PyTorch (CUDA 12.8 support for GPU acceleration)
# Note: Install CPU version if no GPU: torch torchvision torchaudio
--extra-index-url https://download.pytorch.org/whl/cu128
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Transformers for Whisper & Pyannote models
transformers>=4.30.0

# System Monitoring
psutil>=5.9.0

# Audio Device Access
sounddevice>=0.4.6

# ===================================
# Audio Processing Utilities
# ===================================
pydub>=0.25.1
ffmpeg-python>=0.2.0
SpeechRecognition>=3.10.0

# ===================================
# Third-Party API Integrations
# ===================================

# ElevenLabs (Speech-to-Text Alternative)
elevenlabs>=1.0.0

# AssemblyAI (Real-time Transcription)
assemblyai>=0.28.0
websocket-client>=1.6.0

# PyAudio (Microphone & System Audio Recording)
pyaudio>=0.2.13

